Modularize main script
Ticket:3346
