+++ src/aspects/container/container.ts
@@ -1,7 +1,27 @@
 import { Container } from 'inversify';
 import 'reflect-metadata';
 
-// tslint:disable-next-line: no-any
-export function getContainer(...args: any[]) {
-    return new Container(...args);
+class MiBiContainer extends Container {
+
+    constructor(...args: any[]) {
+        super(...args);
+    }
+
+    public bindDependencies(func: Function, dependencies: any[]) {
+        let injections = dependencies.map((dependency) => {
+            return this.get(dependency);
+        });
+        return func.bind(func, ...injections);
+    }
 }
+// tslint:disable-next-line: no-any
+function createContainer(...args: any[]): MiBiContainer {
+    return new MiBiContainer(...args);
+}
+
+
+export {
+    MiBiContainer,
+    createContainer
+};
+
