+++ src/infrastructure/ports.ts
@@ -1,16 +1,17 @@
 // Mail
 
-export { MAIL_TYPES } from './mail/mail.types';
 export { MailService } from './mail/mail.model';
 export { getMailContainerModule } from './mail/mail.module';
+export { MAIL_TYPES } from './mail/mail.types';
 
 // Persistence
 
-export { getPersistenceContainerModule } from './persistence/persistence.module';
-export { UserNotFoundError } from './persistence/model/domain.error';
 export { createDataStore as createParseDataStore } from './persistence/data-store/parse/parse';
+export { UserNotFoundError } from './persistence/model/domain.error';
+export { getPersistenceContainerModule } from './persistence/persistence.module';
 
 export { DataStore } from './persistence/model/data-store.model';
 
-export { initialiseRepository as initialiseCatalogRepository } from './persistence/repositories/catalog.repository';
-export { initialiseRepository as initialiseSearchAliasRepository } from './persistence/repositories/search-alias.repository';
+export { getRepository as initialiseCatalogRepository } from './persistence/repositories/catalog.repository';
+export { getRepository as initialiseSearchAliasRepository } from './persistence/repositories/search-alias.repository';
+
