+++ src/repositorySetup.ts
@@ -1,0 +1,44 @@
+import { configurationService } from './ConfigurationService';
+import { logger } from './aspects';
+import {
+    initialiseCatalogRepository,
+    initialiseSearchAliasRepository
+} from './infrastructure/ports';
+import {
+    DataStoreConfiguration
+} from './main.model';
+
+
+export async function initialiseRepositories() {
+
+    const dataStoreConfig: DataStoreConfiguration =
+        configurationService.getDataStoreConfiguration();
+
+
+
+    const catalogRepository = await initialiseCatalogRepository(
+        dataStoreConfig.dataDir
+    ).catch((error: Error) => {
+        logger.error(
+            `Failed to initialize Catalog Repository. error=${String(error)}`
+        );
+        throw error;
+    });
+
+    const searchAliasRepository = await initialiseSearchAliasRepository(
+        dataStoreConfig.dataDir
+    ).catch((error: Error) => {
+        logger.error(
+            `Failed to initialize Search Alias Repository. error=${String(
+                error
+            )}`
+        );
+        throw error;
+    });
+
+    return {
+        searchAliasRepository, catalogRepository
+    }
+
+}
+
