+++ src/ui/cloud/parseKeys.ts
@@ -1,0 +1,3 @@
+export const CloudFunctionKey = {
+    getInstitutionsList: 'getInstitutionsList'
+};
