+++ src/ui/server/controllers/__tests__/users.controller.spec.ts
@@ -1,24 +1,24 @@
-import { getContainer } from '../../../../aspects/container/container';
-var mockReq = require('mock-express-request');
-var mockRes = require('mock-express-response');
+import { Container } from 'inversify';
+import { rebindMocks } from '../../../../__mocks__/util';
+import { APPLICATION_TYPES } from '../../../../app/application.types';
+import { getMockLoginService } from '../../../../app/authentication/application/__mocks__/login.service';
 import {
     AuthorizationError,
     getApplicationContainerModule
 } from '../../../../app/ports';
-import { UsersController } from '../../model/controller.model';
-import { Container } from 'inversify';
-import { getServerContainerModule } from '../../server.module';
+import { createContainer } from '../../../../aspects/container/container';
 import { mockPersistenceContainerModule } from '../../../../infrastructure/persistence/__mocks__/persistence-mock.module';
+import { UsersController } from '../../model/controller.model';
+import { getServerContainerModule } from '../../server.module';
 import { SERVER_TYPES } from '../../server.types';
-import { rebindMocks } from '../../../../__mocks__/util';
-import { getMockLoginService } from '../../../../app/authentication/application/__mocks__/login.service';
-import { APPLICATION_TYPES } from '../../../../app/application.types';
+var mockReq = require('mock-express-request');
+var mockRes = require('mock-express-response');
 
 describe('Login controller', () => {
     let controller: UsersController;
     let container: Container | null;
     beforeEach(() => {
-        container = getContainer();
+        container = createContainer();
         container.load(
             getServerContainerModule({
                 port: 1,
