+++ src/ui/server/model/controller.model.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 
-export interface Controller {}
+export interface Controller { }
 
 export interface ControllerFactory {
     // tslint:disable-next-line: no-any
